requirements.txt
requests==2.28.2
beautifulsoup4==4.11.2
google-api-python-client==2.70.0
google-auth-oauthlib==0.8.0
openai==0.27.8
schedule==1.1.0
python-dotenv==1.0.0
